requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.8
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==4.1.13
django-cache-url==3.4.5
djongo==1.3.6
dnspython==2.6.1
environs==11.0.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
idna==3.7
importlib_metadata==8.2.0
joblib==1.4.2
keras==3.4.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
protobuf==4.25.4
Pygments==2.18.0
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sqlparse==0.2.4
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zemberek-python==0.2.3
zipp==3.19.2